don't try to dispatch to UNIVERSAL::import
When installing an import, we try to wrap an existing method if it
exists. This may not be a very reasonable thing to do. All of the
original arguments are passed along, including the ones handled by this
module, so the parent probably won't handle them correctly.

Future versions of perl are intending to have a real UNIVERSAL::import
method, rather than a magical special case. And this method will throw
errors if passed arguments, since it means they were not handled. This
conflicts with the behavior of this module, which passes along the
import arguments.

We can detect if the existing import method is UNIVERSAL::import, and
avoid calling it.